requirements.txt
streamlit
python-dotenv
google-generativeai>=0.8.0
streamlit-mic-recorder